requirements.txt
streamlit>=1.28.0
python-docx>=0.8.11
langchain>=0.0.300
chromadb>=0.4.0
sentence-transformers>=2.2.0
google-generativeai>=0.3.0
docxcompose>=1.3.0
pandas>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
